+++ data/scripts/talkactions/player/chain_system.lua
@@ -10,6 +10,11 @@ function feature.onSay(player, words, param)
 		return true
 	end
 
+	if configManager.getBoolean(configKeys.VIP_SYSTEM_ENABLED) and configManager.getBoolean(configKeys.CHAIN_SYSTEM_VIP_ONLY) and not player:isVip() then
+		player:sendCancelMessage("You need to be VIP to use this command!")
+		return true
+	end
+
 	if not table.contains(validValues, param) then
 		local validValuesStr = table.concat(validValues, "/")
 		player:sendTextMessage(MESSAGE_EVENT_ADVANCE, "Invalid param specified. Usage: !chain [" .. validValuesStr .. "]")