+++ src/config/configmanager.cpp
@@ -168,6 +168,7 @@ bool ConfigManager::load() {
 	loadBoolConfig(L, HALF_LOSS_EXP, "halfLossExp", true);
 	loadBoolConfig(L, HALF_LOSS_SKILL, "halfLossSkill", true);
 	loadBoolConfig(L, HALF_LOSS_MAGIC, "halfLossMagicLevel", true);
+	loadBoolConfig(L, CHAIN_SYSTEM_VIP_ONLY, "chainSystemVipOnly", false);
 
 	loadFloatConfig(L, BESTIARY_RATE_CHARM_SHOP_PRICE, "bestiaryRateCharmShopPrice", 1.0);
 	loadFloatConfig(L, COMBAT_CHAIN_SKILL_FORMULA_AXE, "combatChainSkillFormulaAxe", 0.9);